+++ nemo/headers/be/interface/Rect.h
@@ -46,18 +46,18 @@
 // BRect class -----------------------------------------------------------------
 class BRect {
 public:
-	float	left;
-	float	top;
-	float	right;
-	float	bottom;
+	int32	left;
+	int32	top;
+	int32	right;
+	int32	bottom;
 
 	BRect();
 	BRect(const BRect &r);
-	BRect(float l, float t, float r, float b);
+	BRect(int32 l, int32 t, int32 r, int32 b);
 	BRect(BPoint lt, BPoint rb);
 
 	BRect	&operator=(const BRect &r);
-	void	Set(float l, float t, float r, float b);
+	void	Set(int32 l, int32 t, int32 r, int32 b);
 
 	void	PrintToStream() const;
 
@@ -73,25 +73,25 @@ public:
 
 	// transformation
 	void	InsetBy(BPoint p);
-	void	InsetBy(float dx, float dy);
+	void	InsetBy(int32 dx, int32 dy);
 	void	OffsetBy(BPoint p);
-	void	OffsetBy(float dx, float dy);
+	void	OffsetBy(int32 dx, int32 dy);
 	void	OffsetTo(BPoint p);
-	void	OffsetTo(float x, float y);
+	void	OffsetTo(int32 x, int32 y);
 
 	// expression transformations
 	BRect &	InsetBySelf(BPoint);
-	BRect &	InsetBySelf(float dx, float dy);
+	BRect &	InsetBySelf(int32 dx, int32 dy);
 	BRect	InsetByCopy(BPoint);
-	BRect	InsetByCopy(float dx, float dy);
+	BRect	InsetByCopy(int32 dx, int32 dy);
 	BRect &	OffsetBySelf(BPoint);
-	BRect &	OffsetBySelf(float dx, float dy);
+	BRect &	OffsetBySelf(int32 dx, int32 dy);
 	BRect	OffsetByCopy(BPoint);
-	BRect	OffsetByCopy(float dx, float dy);
+	BRect	OffsetByCopy(int32 dx, int32 dy);
 	BRect &	OffsetToSelf(BPoint);
-	BRect &	OffsetToSelf(float dx, float dy);
+	BRect &	OffsetToSelf(int32 dx, int32 dy);
 	BRect	OffsetToCopy(BPoint);
-	BRect	OffsetToCopy(float dx, float dy);
+	BRect	OffsetToCopy(int32 dx, int32 dy);
 
 	// comparison
 	bool	operator==(BRect r) const;
@@ -103,9 +103,9 @@ public:
 
 	bool	Intersects(BRect r) const;
 	bool	IsValid() const;
-	float	Width() const;
+	int32		Width() const;
 	int32	IntegerWidth() const;
-	float	Height() const;
+	int32		Height() const;
 	int32	IntegerHeight() const;
 	bool	Contains(BPoint p) const;
 	bool	Contains(BRect r) const;
@@ -141,7 +141,7 @@ inline BRect::BRect()
 	bottom = right = -1;
 }
 
-inline BRect::BRect(float l, float t, float r, float b)
+inline BRect::BRect(int32 l, int32 t, int32 r, int32 b)
 {
 	left = l;
 	top = t;
@@ -174,7 +174,7 @@ inline BRect &BRect::operator=(const BRect& from)
 	return *this;
 }
 
-inline void BRect::Set(float l, float t, float r, float b)
+inline void BRect::Set(int32 l, int32 t, int32 r, int32 b)
 {
 	left = l;
 	top = t;
@@ -195,7 +195,7 @@ inline int32 BRect::IntegerWidth() const
 	return((int32)ceil(right - left));
 }
 
-inline float BRect::Width() const
+inline int32 BRect::Width() const
 {
 	return(right - left);
 }
@@ -205,7 +205,7 @@ inline int32 BRect::IntegerHeight() const
 	return((int32)ceil(bottom - top));
 }
 
-inline float BRect::Height() const
+inline int32 BRect::Height() const
 {
 	return(bottom - top);
 }
